requirements.txt
pytest
requests
faker
ipdb
python-environ
allure-pytest
jsonpath-rw
pyhamcrest
